Improve ad detection, fix cross-fading in restreamer
Analyzer buffers MFCCs instead of frames, then run prediction on running
buffer, so it predicts more often. However, there is still about 1.5sec
delay in prediction because the classifer needs full data frame of Ads
to consider it as Ads, and full frame of music to consider it as music.
I think it can only be fixed with shorter data frame.

Restreamer has better cross-fading now. It always use Ads frames to run
crossfade, so it might be required to pad ad with silence to avoid
cutting advertisement content.

+++ classifier/src/types.rs
@@ -1,4 +1,4 @@
-pub type Data = ndarray::Array4<f64>;
+pub type Data = ndarray::Array4<f32>;
 pub type PredictedLabels = ndarray::Array2<f32>;
 pub type Labels = ndarray::Array1<u32>;
